+++ crates/svelte_analyze/src/types/mod.rs
@@ -3,6 +3,17 @@ pub mod script;
 pub mod node_table;
 pub(crate) mod markers;
 
-pub use data::*;
-pub use script::*;
-pub use node_table::*;
+pub use data::{
+    AnalysisData, AsyncStmtMeta, AwaitBindingData, AwaitBindingInfo, BindSemanticsData,
+    BlockerData, ClassDirectiveInfo, CodegenView, ComponentBindMode, ComponentPropInfo,
+    ComponentPropKind, ConstTagData, ContentStrategy, DebugTagData, DestructureKind,
+    EachBlockData, ElementFlags, EventHandlerMode, ExprDeps, ExprSite, ExpressionInfo,
+    ExpressionKind, FragmentData, FragmentItem, FragmentKey, IgnoreData, LoweredFragment,
+    LoweredTextPart, PropAnalysis, PropsAnalysis, RenderTagArgPlan, RenderTagCalleeMode,
+    RenderTagPlan, RuntimePlan, SnippetData, TitleElementData,
+};
+pub use script::{
+    DeclarationInfo, DeclarationKind, ExportInfo, PropInfo, PropsDeclaration, RuneKind,
+    ScriptInfo,
+};
+pub use node_table::{NodeBitSet, NodeTable};
