+++ crates/svelte_codegen_client/src/script/location.rs
@@ -1,0 +1,22 @@
+/// Sanitize a filename for use in trace labels by inserting a zero-width space
+/// after each `/` to prevent devtools from treating it as a clickable link.
+pub(crate) fn sanitize_location(filename: &str) -> String {
+    filename.replace('/', "/\u{200b}")
+}
+
+/// Compute 1-based line and column from source text and byte offset.
+pub(crate) fn compute_line_col(source: &str, offset: u32) -> (usize, usize) {
+    let offset = offset as usize;
+    let bytes = source.as_bytes();
+    let mut line = 1;
+    let mut col = 0;
+    for i in 0..offset.min(bytes.len()) {
+        if bytes[i] == b'\n' {
+            line += 1;
+            col = 0;
+        } else {
+            col += 1;
+        }
+    }
+    (line, col)
+}
